move history.rs to execution mod
