+++ scylla/src/execution/iterator.rs
@@ -25,6 +25,7 @@ use crate::transport::errors::QueryError;
 use crate::cluster::ClusterData;
 use crate::cluster::{Node, NodeRef};
 use crate::connection::{Connection, NonErrorQueryResponse, QueryResponse};
+use crate::execution::history::{self, HistoryListener};
 use crate::execution::load_balancing::{self, RoutingInfo};
 use crate::execution::retries::{QueryInfo, RetryDecision, RetrySession};
 use crate::frame::{
@@ -34,7 +35,6 @@ use crate::frame::{
     },
     value::SerializedValues,
 };
-use crate::history::{self, HistoryListener};
 use crate::statement::Consistency;
 use crate::statement::{prepared_statement::PreparedStatement, query::Query};
 use crate::transport::metrics::Metrics;
