+++ scylla/src/statement/batch.rs
@@ -1,9 +1,9 @@
 use std::borrow::Cow;
 use std::sync::Arc;
 
+use crate::execution::history::HistoryListener;
 use crate::execution::retries::RetryPolicy;
 use crate::execution::ExecutionProfileHandle;
-use crate::history::HistoryListener;
 use crate::statement::{prepared_statement::PreparedStatement, query::Query};
 
 use super::StatementConfig;
